requirements.txt
tweepy
mysql-python
